Fixed signals bugs and UI enhancements
Updates the process progress calculation to display as a percentage.

Enhances the UI by:
- Adding a window title.
- Including a processor image.
- Improving input validation and feedback for adding processes.
- Adjusting column widths in the process table.
- Disabling modifications to the scheduler selection after a process is added.

The signal connections are simplified by removing redundant explicit connections.

Adds a resource file for including images.

+++ frontend/processProgress.cpp
@@ -6,7 +6,7 @@ void ProgressBarDelegate::paint(QPainter *painter, const QStyleOptionViewItem &o
                                 const QModelIndex &index) const
 {
     // Get the progress value (should be 0-100)
-    int progress = index.data(Qt::DisplayRole).toInt();
+    float progress = index.data(Qt::DisplayRole).toFloat();
 
     if (progress >= 0 && progress <= 100)
     {
